Changing the way errors are reported to the parser to be return code based instead of panicking.
Turns out that PyPy is deeply (and not unreasonably) upset by us panicking through its call stack.
The only real solution is just not to do it and to return errors instead, although it's quite awkward in a number of places
to find ways to signal errors correctly. Python can then reraise them as exceptions which are caught as normal.

This implies that it's now possible for BUILD code to catch such exceptions and handle them, although
I don't particularly recommend that as a strategy for writing BUILD files.
The new exceptions are currently instances of ParseError which is a new global, but we don't make
a specific guarantee that other exceptions won't be raised either.