Reasonably principled solution to the plz run situation with java_binary rules. (#96)
Invoke them with java -jar when doing $(exe //target), but only when not tagged as self_executable so we don't
overwrite jvm_args etc. This also covers plz run, but needs some minor changes for the builtin rules to be
a bit more principled about which things they find via $(exe ) and which via $(location ).