Initial cross-compiling support.
Essentially we identify cases when we require a different architecture to the host and create duplicates of targets for that arch.
They live in different directories under plz-out etc.
There are two ways that can happen; either on the command line there's a flag for the arch to target, or targets can specify that they always target a specific arch
(for example, a target building a rkt container might always be Linux).

The actual support is fairly lightweight; BUILD files are not reparsed so swapping on CONFIG.OS / CONFIG.ARCH in them is a little dangerous. Those are exposed as variables
when the target builds. That's sufficient for relatively simple cases like Go (where we just need to set GOOS / GOARCH to cross-compile), Java (where it's really just
maven_jar rules with native dependencies to worry about) and possibly Python (depending on the pip situation, might need to suppress .pyc generation too).
It's unlikely to be very satisfying for C, but I'm not sure I want to build complete support there right now.