Fix go_get with dependencies when in declared in repo root
When in the root, building a go_get with a dependency, also in the root,
fails, because PKG="" and PKG_DIR == "." and with
GOPATH=$TMP_DIR/TMP_DIR/$PKG, both the deps and the buildable get built
in the same folder. Thus, the usual mechanism of filtering out the dep
files (find $PKG_DIR and delete the equivalent in the $TMP_DIR) does not
work because we cannot distinguish these anymore. Further more, using
"." in the sed command gets interpreted as a wildcard so the sed .. | rm
currently does not delete anything. The end result is that the buildable
produces .a outputs which are also produced by its dependencies and plz
detects that multiple targets output the same file.

The PR saves the existing outputs in the PKG_DIR before performing go
install and uses this for the clean up. It also special cases the "." to
make the sed command work as desired.