Ensure all go_get dependencies are used and filtered
Currently only dependencies in the top level folder and the current
go_get target's folder will be used in the GOPATH and considered for
deletion. Anything else specified as a dependency will not be; thus one
might end up re-downloading a dependency with a different version.

This change makes sure any go_get dependency is included in the GOPATH
regardless of location and also makes sure their outputs are excluded
from the go_get being built