Fix incorrect logic in test runs calculation.
Should be multiplication instead of division; basically if you run `plz test //blah` and //blah is flaky, it runs up to three times until it gets one success.
If you run `plz test -n 5 //blah` then previously it would run until it got two successes (because the 5 was divided by the flakiness).
Now it will run it at least 5 times and up to 15, which seems much more in line with other cases.