Statically linked Python interpreter (#255)
* starting on static interpreter

* Rename functions to prevent name clashes

* hack it in statically

* Remove portable PyPy stuff. Disable PYTHONHOME

* Don't import site

* Set PYTHONHOME to current executable dir

* attempting to make _cffi_backend available

* genericising via config

* vendorising bits of path to stop the madness

* defaultdict is no longer available

* fixing up some tests

* fixing up remote test

i always knew this thing was going to be a problem...

* clean up old stuff

* various documentation updates and whatnot

* clean up bootstrap

* breaking down the lib_py stuff

* fix these tests again

* Move libffi up to top level, preload with dlopen so it doesn't matter where it is.

Easier to handle when it's not in a subdir (I'm not sure the update code will handle
that correctly, for example; nor does the tarball builtin rule)

* updating Docker image

Tagging temporary label since this branch needs to test against it. Not quite sure if this is the best way of managing it...

* Add some special casing for OSX which *might* help?

* sort out inevitable libffi issue, try to improve pkg-config situation a bit

pkg-config seems like it's going to become untenable though. Unsure what a better solution is unfortunately...

* update pkg-config, pypy now unnecessary on osx

* mark test as flaky, have seen it fail other times too

* attempt better OSX search

weirdly --static doesn't seem to have the correct effect on pkg-config - but am guessing a bit with the machine I've got available

* don't need python2, we already have it \o/

* add pypy back, is probably the easiest way to get cffi...

* Get rid of mad library scheme, have persuaded linker to play ball

* possibly this works?

* only preload the .so on linux

* fix after rebase

* Dimitar's comments