Experimental Go-native BUILD file interpreter.
This is somewhat more limited than the current Python-based implementation; some
changes to existing build rules will be needed to be compatible, and hence it is
not yet the default. It is used during Please's bootstrap and hence is fully
self-hosting.

It's generally several times faster than the incumbent, partly due to removing the
cgo<->Python overhead and the GIL.

Notably, it supports and enforces type annotations on functions. This is an
alternative to the relatively crude codegen solution used before (although that
is still retained for the older parse engines).

Several bugs remain but their discovery is left as an exercise for the reader.
The adventurous can try it by passing `-o parse.engine:asp` (or altering their
.plzconfig similarly).