Redoing subinclude deferral process
Instead of the old backoff-and-retry process which had some awkward consequences, this now causes goroutines to hang up synchronously
waiting for things to be done. This cleans the parse code a lot, admittedly requiring a little more fiddling in thread management.

Some care is needed because we have a limited supply of workers and can't block them forever, but also spawning a large number of
goroutines for parse tasks (i.e. making them all parallel and asynchronous) results in significantly lowered performance due to
(presumably) too much contention.

This is probably required for both better Bazel compatibility and cross-compiling (via a strange and awkward chain of semi-dependencies that are best not asked about).