Reworking thread-replacement strategy in favour of an adjustable pool.
This remains pretty fiddly due to the most obvious approach (running all parse actions in parallel goroutines) being
rather slower, presumably due to excessive contention. On the bright side it requires less cooperation through
the rest of the system than the last version and resolves some of the bugs with it.