Making Java a soft build-time dependency
Basically we now have an option on downloading prebuilt junit_runner / javac_worker jars.
Bootstrap script selects appropriately based on whether javac is present or not. This avoids
the slightly awkward description in the README of not needing other dependencies but
still needing Java to build (although python turns out to be a little more needed than I
realised, but I think that is solvable too).