Cross-compiling support
This is definitely rudimentary in terms of actual language support but the primitives are there in the build system.
This is triggered either by the -a / --arch flag or a target can specifically request an architecture of a dependency using the subrepo syntax (@linux_x86//pkg:target).
Cross-compiled targets will read an appropriate config file and so can be configured to use different tools.

Doesn't really do anything out of the box for most languages right now though.