Change JUnit Runner output to be equivalent to Python test output (#321)
The Python XmlRunner captures a lot richer information and it would be nice to surface that to either the collation step or do away with the collation entirely - CIs like Jenkins generally do a good job of parsing the multiple TEST-foo.xml files that tools like maven-surefire generate.

This PR alters the JUnit Runner to output a file called TEST-className.xml based on the maven surefire xml format. This handles skipped tests, and the difference between errors and failures better.

It has been refactored to separate coverage runs from test runs, and running and reporting are now separated too, making it possible to unit test both in isolation. The TestMain.java remains the entrypoint to the susbsytem, orchestrating which classes to run before handing them off to the runners and reporters.