Replace Go BuildId with something that does not change (#373)
The BuildId is based on current directories, environment variables and so on. This means that the same build on two different machines will only have the same build id if they happen to have the same directory structure, and will have different ones otherwise. This is used as a cache key.

As Please handles caching at a different level, we overwrite the BuildId with "BuiltWithPleaseBuild/BuiltWithPleaseBuild" to ensure it is consistent across builds.