Added functionalities for Call type hoverring (#433)
I needed the Pos and EndPos for the IdentExpr as well, so I added them on to the grammar.go. Not so surprisingly, it cleaned up that function parseIdentExpr.

Most of the puzzle pieces for hovering should be done. I have used that in function call type of IdentStatement. The rest of it should be only fitting the pieces together. (Unless I changed my mind again)

Added functionality in analyzer.go for obtaining info on BuildLabels(Analyzer.BuildLabelFromString()).