Generate IntelliJ project files for Java (and Scala) targets (#437)
The basic idea is to map java_library and maven_jar targets to IntelliJ modules and libraries, and create the chain of dependencies between them. Libraries cannot depend on anything so there needs to be intermediate modules created for each third_party maven_jar target, which can then depend on other modules or libraries as desired.

This introduces a set of new magic labels (notably package_prefix which tells IntelliJ to pretend the module is rooted at that location, meaning it doesn't need to traverse the tree). Other tools like Pants handle this by forcing all Java code to be under a set of predefined paths like src/java from which it can determine the package structure.