added workspace store and fs methods handlers (#446)
With both completion and hover, we will need a way to track changes in document-in-editing.

I have added a workspaceStore struct to save documents in editing
Also added file system requests handlers to update the document store.
Some of the existing file reading will need to be changed, because we are currently only reading the saved files, which is not going to work.
I'm not sure how deleting line request are being send for other editors, I noticed a bug with deleting line in the intellij LSP plugin, (handled in workspace_store.go, line 95-98), however, after testing it out with various cases, it's still not exactly ideal. Let me know if you have any thoughts on this.