Add native git_*() commands to plz... (#506)
* build_rule: add extra_hash_data attribute

Used by `go_get()` and `go_test()`'s `definitions` arguments as a way
of updating the filesystem hash for targets based on their linker flags.

Fix 2/3 from #501

* git_show(): add new function that can be used to extract various git metadata

git_show() can be used for getting the UNIX time of the last commit,
commit hashes, PGP information, or information about the author.  See
`git-show(1)` for additional details.  The verb accepted by
`git_show()` is a subset of what `git show` supports.

* fs/hash: back out using PathHasher as a way of hashing outputs

* src/build: use RuleHash() to hash extra_hash_data

* src/build: fix up incremental tests for ExtrHashData

* builtins: add git_state()

Returns "clean" or "dirty" based on the state of the repository.

* builtins: add git_commit()

Returns the git commit SHA to the caller.  If the argument `short=False`
is passed, the full git sha is returned (40char vs 8char).

* builtins: add git_branch() support

Returns the name of the current branch.  If short=False, return the
full symbolic ref for a given branch.

* Make buildifier happy

* exec: only expose the full sha out of git_commit()

User-specified configuration files overriding the definition of
short undermines any reproducitbility guarantees.  Users of this
function call should call `git_commit()[0:8]` to truncate a sha
to an acceptable length.

Also remove any format that may be impaired by user locale or settings

* exec: clean up doExec()

* docs: move git_* under misc

* backout a mis-merge from gh-501b