Improve git/exec error msg (#510)
* exec(): Cache the call to exec.LookPath() to resolve the abspath of commands

This may also help aid in improving error messages in CI.

* git_branch(): fix branch detection for detached heads

In CI systems its common to checkout a git sha as a detached head.
git_branc() doesn't like that very much.  Use the last element of the
ref name as our approximate git branch.

* git_branch(): stuff in a `-q` arg for development

Local diffs are not welcome in `git show` output here, we just want the
one line git ref.