exec: fix race in promise cache
After the move to support `git_branc()` on detached HEADs, it became
trivial to reproduce a race condition in the promises code where
concurrent access to the same exec() command were missing the sync.Cond
Broadcast.  This was needlessly complicated (the last chunk of code I
wrote before this patch used condition variables, so it was top of mind
at the time).  Use of a sync.WaitGroup is actually much more correct in
this case.  Also, use `sync.Map` and introduce a negative cache for
commands that are expected to fail (such as those used in `git_branch()`).