Refactoring display stuff out of plz package
Been noising about this for some time. It's a nicer design to be able to use that package to build things without having to configure anything about the display; also the display code shouldn't be responsible for
deciding whether the build counts as successful or not.
Slims the package interface down a bit since 80% of the properties in InitOpts related to display (unsurprisingly since the rest were already on State because other things would need them).

Removed VisibilityParse as part of this process. We recently discovered that it can cause spurious failures if there's a visibility label that corresponds to a directory that no longer exists, and I don't
think it earns its living any more. It dates back to 2017 (pre parser rewrite) when that operation was a lot slower, and anecdotally rarely makes a difference (because most times you really care about
revdeps the target will tend to be publicly visible anyway).