Add cirrus-ci configuration for FreeBSD (#509)
* Add cirrus-ci config file

* suspect that this "darwin" fix might in fact be "anything not Linux"

* add a sweet shield to README

* Exclude several failing tests

Not really sure what is up with them but there is precedent in the no_circle label. Seems more useful to turn it on minus a few tests than spend forever trying to figure those out.

* Protobuf version bump.

Looks like maybe FreeBSD has a newer version (and I don't think we can choose a specific one from ports which might be troublesome eventually but wevs)