Add Go manually on OSX (#554)
Not really sure of all the detail; Homebrew seems to fail with older versions of xcode, but newer versions cause many other things to fail with "no such file" errors. Not sure if this is XCode or Circle or what, but this is quite a bit faster than waiting for Homebrew anyway and these days we don't need many dependencies.