Additional fallback for finding subrepos.
If a subrepo defines a dependency on a subrepo of its own in a package that we have at the top level (as described in #577) then we don't find it (we already do have a fork for when that does not exist at the top level).
This lets the top-level repo override it but if it's not specified we can still find it.