Small go_get improvement. (#608)
Include the last part of the repo in the naming formula for go_get
downloads. As previously written, "golang.org/x/crypto" and
"golang.org/x/sync" could not exist within the same go_get rule with
revisions declared. Omitting revisions caused a fallback to the
slow-path of `git clone`.