Deprecate RPC cache in favour of rex
For now it will continue to work, but we won't release any new versions of the server. Update docs to explain the current situation (which is a bit of a classic "deprecated vs. not ready yet" but better to be clear about that - also the http cache is a pretty valid replacement anyway).