Better synchronisation of updates when build completes
This eliminates an awkward pause where it stopped displaying what was happening but was still waiting for running build actions before terminating.
Happy with most of it, CloseResults() is a little yucky but the owner of the worker threads really needs to manage that. I think it might fix the "send on closed channel" nonsense too...