Fix for jarcat when rules are sandboxed
Getting errors like `panic: unable to copy the file we just wrote (/tmp/jarcat-085478954): mkdir : no such file or directory` from all actions that use it.
This change makes it always use the current directory which is more like what happens when it's not sandboxed (when TMPDIR is the same as the temporary build dir).