Add an option for authenticating using a pre-shared token.
It would be conceptually nice to use the CredentialsFile option on the SDK, but after quite some experimentation I can't find anything able to validate that on the server side.
Unsure what is wrong there... but also we have some use cases involving Istio where we wouldn't necessarily have TLS from the client's perspective so that wouldn't be an option anyway.