go_test(): avoid redefined imports in generated test main files (#1066)
When a go_test() target is defined for a test file belonging to any of
the following packages, `plz build` fails when compiling the test file:

* os
* strings
* testing
* testdeps

This is because the test main file template already imports packages
with these names from the Go standard library before the test package
with an identical name is imported. Alias these imports with the prefix
"_gostdlib_", making it much less likely that WriteTestMain() generates
invalid code because of redefined imports.

Fixes #1065.