Implement retries to the HTTP cache with go-retryablehttp
A number of builds that I was doing rely on a somewhat unreliable
webdav-S3 server bridge, where occasionally timeouts happen or a file
cuts of in transit due to one error or another. Instead of immediately
pretending a file doesn't exist (or never uploading a file), retry
a certain number of times first in a controlled manner.

Closes #1140

Signed-off-by: Robert Xu <me@robxu9.com>