jarcat/unzip: create entire path when creating directory Tar members (#1203)
Following #1200, jarcat/unzip no longer explicitly creates directories
when processing members of Tar archives. This is problematic when
extracting an archive containing an empty directory, since there is no
future action that will cause that directory to be created (e.g. the
creation of intermediate directories for a child file, since there are
none).

Mostly revert #1200, and explicitly create directory Tar members using
os.MkdirAll(), which avoids raising an error when the target path
already exists (which is the problem #1200 intended to solve).