java_test(): handle case where $TOOLS_JUNIT is a real file
TOOLS_JUNIT can sometimes contain a path to a real file as well as the
name of an executable file in PATH. Handle either eventuality by
creating a symlink to $TOOLS_JUNIT if it is a path to a real file, or
creating a symlink to the output of `which $TOOLS_JUNIT` otherwise.

Fixes #1206.