go_get: use entire package names in tag names (#1327)
When generating tag names from package names in the go_get rule, use a
sanitised version of the entire package name rather than the
last-but-one component in the path. This avoids breakage in cases where
multiple packages with the same last-but-one path component are fetched
as part of the same go_get rule (e.g. https://github.com/x/y and
https://github.com/x/z).

Closes #1326.