go_get: allow per-target patches to be defined
In the go_get rule, allow the value of the 'patches' parameter to be a
dict whose keys are target names given in the 'get' parameter and whose
values are the patch(es) that should be applied to that target.
Backwards compatibility is maintained: if 'get' is a list and 'patches'
is a string or list, the patch(es) are applied to all targets in 'get'.

Closes #1332.