go_get: allow per-target package installs to be defined
In the go_get rule, allow the value of the 'install' parameter to be a
dict whose keys are target names given in the 'get' parameter and whose
values are the packages that should be installed from that target.
Backwards compatibility is maintained: if 'get' is a list and 'install'
is a list, Please will still install all packages in 'install' from all
targets in 'get'.

Closes #1334.