go_get: allow per-target paths to be stripped
In the go_get rule, allow the value of the 'strip' parameter to be a
dict whose keys are target names given in the 'get' parameter and whose
values are the paths that should be deleted from that target's directory
before it is built. Backwards compatibility is maintained: if 'get' is a
list and 'strip' is a list, Please will still delete all paths in
'strip' from all targets in 'get' before building them.

Closes #1337.