sandbox: ignore non-existent directories in `SANDBOX_DIRS` (#2041)
`mount()` sets `errno` to `ENOENT` when the target doesn't exist, which
causes the sandbox to error fatally when it is asked to mask a directory
that doesn't exist on the host. Instead, output the existing "Not
mounting over <target> since it isn't a directory" warning that is shown
when the given target is a non-directory file, and continue.

Fixes #2040.