Allow `Python.DefaultInterpreter` to be a build target name
Allow the path to the default Python interpreter to be a Please build
target name, in which case that target will be built and the full path
to its output binary will be embedded in PEX files generated by
`python_binary` and `python_test` targets. This allows `python_binary`
and `python_test` targets to depend on a Python interpreter that is
itself built in the repository without first needing to install the
interpreter externally.