`cc_static_library`: expose `hdrs` to dependent targets
Targets that depend on archives collected by a `cc_static_library`
typically need access to the headers that were used to build those
libraries, but they currently aren't exported to dependent targets when
`cc_static_library` is used purely as a collector for other archives.
Export these headers when `hdrs` is defined but `srcs` isn't.