Make C/C++ library output filenames customisable
There are occasions when a C/C++ library needs to be given a specific
filename, e.g. when it needs to be automatically detected by an
Autoconf-generated configure script, but doing this using the built-in
rules requires either a library target whose name matches the desired
library name or an additional `genrule` target to do the renaming if
this isn't possible.

Add an `out` parameter to the `cc_library` and `cc_static_library` rules
(and their C equivalents) to allow the library filename to be
customised. Similarly to the existing behaviour for `cc_shared_object`,
a sensible default is chosen if no value of `out` is given.