`cc_*`: collect transitive labels when only `hdrs` is defined
Run the pre-build function `_library_transitive_labels` when `hdrs` is
defined for `cc_library` and `cc_object`, and run the pre-build function
`_binary_transitive_labels` when `hdrs` is defined for
`cc_shared_object`. This ensures that the labels attached to the `hdrs`
rules are picked up by the dependent target and therefore that the paths
to the headers are correctly added to the C/C++ compiler's header search
path, even when the `deps`, `includes` and `defines` parameters are not
defined on the dependent target.

Fixes #2251.