Sanitise environment variable values when executing remote actions
When executing remote actions to build targets that have a custom
environment, Please sets up the environment in the remote shell by
executing an `export` command for each environment variable before
executing the build command. The environment variable values aren't
sanitised, so shell constructs such as `$(...)` will be interpreted
(which isn't the case when building locally, since the environment is
set up by `os/exec` instead), and any value containing a double-quote
causes the action to fail altogether.

When constructing the `export` commands, sanitise environment variable
values using [shellescape](https://github.com/alessio/shellescape). This
ensures not only that double-quotes in the value don't cause failures
when building the target remotely, but also that the environment when
building the target remotely is identical to that when building it
locally.

Closes #2407.