Stay on Git branch when using `--since` with `query changes` (#2431)
When used on a Git repository, `plz query changes --since <S>` gets the
commit hash of the current HEAD `C`, checks out the revision `S`, then
checks out `C`. If the repository was on a branch before the command was
run, Please restores the working tree to `C` rather than the branch for
which `C` is the HEAD, leaving the repository in detached HEAD state.

Before checking out `S`, run `git symbolic-ref -q --short HEAD` to
determine the name of the current branch (if any), falling back to `git
rev-parse HEAD` if this fails (which indicates that the working tree is
not on a branch), and use this as the value of `C`. This ensures the
repository is always restored to the state it was in before `plz query
changes` was executed.

Closes #2423.