`java_library`: output JNI headers generated by `javac`
The Java compiler outputs header files for Java sources that use the
Java Native Interface (JNI), but there's currently no way for them to be
consumed by the native-code components that depend on them, preventing
them from being built.

Provide the JNI headers as outputs of `java_library` when `cc_hdrs` is
required. This allows native-code JNI components written in C/C++ to be
built by including the `java_library` target as a dependency.