fix(core/package): Fix dependency resolution before showing warning (#2554)
The condition used to check for the parent of the rule only, this adds
a check for the rule itself before showing the warning.

A better version of this would be to check whether there is a conflict
of files between the two rules but I'm not sufficiently accustomed with
the codebase yet to attempt that endeavor. For now, this PR opens an
escape hatch for those who know what they're doing to at least silence
the warning.