Exclude targets in plugin subrepos from `plz query changes`
Changing a `plugin_repo()` target causes all of the plugin subrepo's
child targets to be outputted by `plz query changes`. This is
problematic because it isn't necessarily desirable (or even possible,
depending on the environment) to build every single target in the plugin
subrepo.

Exclude all plugin subrepo child targets from the output of `plz query
changes` by default, and add an `--include_plugins` option to revert to
the previous behaviour if desired.