`github_repo`: strip `v` prefix if `revision` is any version number
`github_repo` currently strips the `v` prefix from the file names in the
zip file downloaded from GitHub if `revision` is a semver, since GitHub
silently does the same if asked to archive the code for semver-like
tags. It looks like this logic extends to more than just semvers,
though - the same behaviour has been observed for strings that are
simple dotted version numbers with any number of components (e.g.
`v1.1.10.1`). Account for this in the prefix-stripping logic.

Fixes #2829.