Add `pop` method to dict built-in type
```
dict.pop(key[, default])
```

This method removes the given key from the dictionary and returns its
value if it existed, or the given default if it didn't. The semantics
are similar to Python's and Starlark's in that `dict.pop(key)` causes
an error to be raised when `key` does not exist, similarly to
`dict[key]`.