`format`: only sort `tag` parameter below `name` for built-in rules
Turns out `tag` is quite a common parameter name for custom build rules.
We only want it to sort below `name` in cases where the parameter's
value is actually part of the target name, i.e. for the Please built-ins
`build_rule` and `filegroup`. In all other cases, it should sort
alphabetically with the unprioritised parameters.