Return `None` for empty optional plugin config fields
Specifying an empty value for an optional plugin configuration field
causes the field to be set to the type's default value (`""` for
strings, `0` for integers, `False` for booleans) from the perspective of
ASP. This makes it impossible to tell whether an optional field was
actually assigned a value inside a `BUILD` file.

For optional fields, return `None` if no value is assigned.