golangci-lint: disable staticcheck's error formatting checks (#3385)
These checks relate to the capitalisation of leading letters and the
existence of trailing punctuation characters in `error`s. While we ought
to be following the Go conventions for these, there are too many
instances (84) of them right now, and the affected errors are so deeply
nested that it's difficult to identify the impact that reformatting them
will have on Please's output. Disable the issue for now.