Register architecture subrepos before parsing parent package
`maybeParseSubrepoPackage` checks whether the package containing a
subrepo exists as a precondition of registering the subrepo. Because
this involves parsing the package's `BUILD` file, Please's built-in
architecture subrepos - implicitly defined at the top level - are not
registered unless the repo has a top-level `BUILD` file, even an empty
one.

Initially attempt to register a subrepo as an architecture subrepo
before falling back to parsing the parent package's `BUILD` file and
searching it for the required subrepo.

Fixes #3408.