Remove embedded binary from cross-compilation test
This test predates the migration to plugins - `c_embed_binary` was added
to cc-rules in v0.1.0 (although it will be removed in v0.5.0 due to lack
of arm64 compatibility). The embedding part of the test isn't necessary
to demonstrate that cross-compilation works, so remove it. If
`c_embed_binary` is re-added to cc-rules later, this test can import the
`c_embed_binary` provided by the plugin.