Implement run-time dependencies for binary targets (#3451)
Add a `runtime_deps` parameter, allowing for the declaration of run-time
dependencies, to `build_rule` and several of the built-in build
definitions that permit binary outputs.

The semantics of `runtime_deps` are a combination of the semantics of
`deps` and `data`. If the output of a `build_rule` is marked as binary,
targets listed in `runtime_deps` are:

- guaranteed to have been built before the dependent target runs;
- copied into build and test environments alongside the dependent
  target.

Like build-time dependencies declared with `deps`, run-time dependencies
are computed transitively; i.e., if target A has a run-time dependency
on target B which in turn has a run-time dependency on target C, targets
B and C will be built before `plz run`ning target A, and the outputs of
targets B and C will be copied into build environments (or test
environments, if target A is a test) alongside the outputs of target A.
Unlike build-time dependencies, downward searches for transitive
run-time dependencies are not blocked by the `output_is_complete`
parameter.