Add `maxdepth` parameter to `get_labels` function
This is used in conjunction with the `transitive` parameter; when
`transitive` is true, `maxdepth` can be used to control the maximum
depth of the recursion. The default is -1, meaning "no limit". The main
use case is to help limit the collection of `cc:inc:` labels to direct
dependencies in the cc-rules plugin, since headers used by (library)
dependencies shouldn't be required at compilation time.

`transitive=True, maxdepth=0` is equivalent to `transitive=False`, so
actually there's no need for `transitive` to exist as a parameter at all
any more, although it remains for backwards compatibility. It could be
removed in a future major release.